refactor: common-lib으로 예외 처리 및 API 응답 구조 이동
