--- common-lib/src/main/java/com/realtime/commonlib/CommonLibApplication.java
@@ -1,13 +1,0 @@
-package com.realtime.commonlib;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CommonLibApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CommonLibApplication.class, args);
-    }
-
-}
