+++ common-lib/src/main/java/com/realtime/commonlib/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package com.realtime.serving.global.exception;
+package com.realtime.commonlib.exception;
 
 import lombok.Getter;
