--- common-lib/src/test/java/com/realtime/commonlib/CommonLibApplicationTests.java
@@ -1,13 +1,0 @@
-package com.realtime.commonlib;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CommonLibApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
