+++ serving-system/src/test/java/com/realtime/serving/application/service/KeywordApplicationServiceTest.java
@@ -5,13 +5,13 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.verify;
 
+import com.realtime.commonlib.exception.BusinessException;
+import com.realtime.commonlib.exception.ErrorCode;
 import com.realtime.serving.TestAssertions;
 import com.realtime.serving.TestDataFactory;
 import com.realtime.serving.domain.keyword.entity.Keyword;
 import com.realtime.serving.domain.keyword.entity.TrendStatus;
 import com.realtime.serving.domain.keyword.repository.KeywordRepository;
-import com.realtime.serving.global.exception.BusinessException;
-import com.realtime.serving.global.exception.ErrorCode;
 import com.realtime.serving.presentation.rest.keyword.dto.KeywordCreateRequest;
 import com.realtime.serving.presentation.rest.keyword.dto.KeywordResponse;
 import java.util.List;
